% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesLearnSU.R
\docType{package}
\name{BayesLearnSU}
\alias{BayesLearnSU}
\title{BayesLearnSU: a package for the Bayesian learning course at SU}
\description{
This package contains the data sets used in the course Bayesian learning at
Stockholm university. The package uses lazy loading, which means that the 
data sets are not loaded themselves into the environment, but they will be
on first access.
}
\details{
This means that while (for example) tempLinkoping does not show up in the
environment, you can still use it. (As an example, after loading the package
try "head(tempLinkoping)" or "mean(tempLinkoping$temp)".)

For a list of all available data sets, use help(, "BayesLearnSU").
}
