% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_normal_approximation.R
\name{beta_normal_approximation}
\alias{beta_normal_approximation}
\title{Plotting the normal approximation of the Beta distribution}
\usage{
beta_normal_approximation(n, s, alpha = 1, beta = 1)
}
\arguments{
\item{n}{Number of trials in the Beta}

\item{s}{Number of successen in the Beta}

\item{alpha}{Prior hyperparameter alpha}

\item{beta}{Prior hyperparameter beta}
}
\description{
Plots a Beta distribution together with the corresponding 
  normal approximation.
}
