% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebayBiddersData.R
\docType{data}
\name{ebayBiddersData}
\alias{ebayBiddersData}
\title{Coin auctions from ebay}
\format{
A data frame with 1000 rows and 10 variables:
\describe{
  \item{nBids}{The number of bids in the auction}
  \item{const}{The intercept}
  \item{powerSeller}{Is the seller selling large columes on ebay?}
  \item{verifyID}{Is the seller verified by ebay?}
  \item{sealed}{Was the coin sold sealed in a never opened envelope?}
  \item{minBlem}{Did the coin have a minor blemish/defect?}
  \item{majBlem}{Did the coin have a major blemish/defect?}
  \item{largeNeg}{Did the seller have a lot of negative feedback from
  customers?}
  \item{logBook}{Logarithm of the coins book-value according to expert sellers. Standardized.}
  \item{minBidShare}{A variable that measures the ratio of the minimum
  selling price and the book value. Standardized.}
  ...
}
}
\usage{
data(ebayBiddersData)
}
\description{
A data set of 1000 auctions of coins on ebay.
}
\keyword{datasets}
