% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Rainfall at Snoqualmie Falls, Washington}
\format{
A data frame with 6920 rows and 1 variables:
\describe{
  \item{precipitation}{Daily precipitation in 0.01 inches, excluding zeroes.}
  ...
}
}
\usage{
data(rainfall)
}
\description{
A dataset of daily records of precipitation in 0.01 inches, with zeroes
excluded. The data was collected between 1948 and 1984 at Snoqualmie
Falls, Washington.
}
\keyword{datasets}
