% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesLearnSU.R
\docType{package}
\name{BayesLearnSU}
\alias{BayesLearnSU}
\title{BayesLearnSU: a package for the Bayesian learning course at SU}
\description{
This package contains the data sets used in the course Bayesian learning at
Stockholm university.
}
\details{
For a list of all available data sets, use help(, "BayesLearnSU").
}
